+++ spec/support/repositories/sequel/post_repository.rb
@@ -4,6 +4,8 @@ class PostRepositorySequel
   for_entity Post
   default_provider :Sequel
 
-  map_attribute :author_name, :to => 'Author#name', :via => :author_id
-  map_attribute :author_bio,  :to => 'Author#bio'
+  group :provider => :Sequel, :via => :author_id do
+    map_attribute :author_name, :to => 'Author#name'
+    map_attribute :author_bio,  :to => 'Author#bio'
+  end
 end