+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## master
 
+- New API: Introduced `Datamappify.config` for configuring default behaviours.
 - API change: `where` now replaces `find` for multiple items.
-- Introduced `Datamappify.config` for configuring default behaviours.
 - API change: `map_attribute` now takes `:to` and `:provider` instead of a string for specifying data source.
+- New API: Introduced `group` for grouping `map_attribute`.
 
 ## 0.53.2 [2013-07-18]